Update student context
